requirements.txt
Django==2.1.4
djangorestframework==3.9.0
django-model-utils==3.1.2
django-jet==1.0.8
celery==4.2.1
redis==3.0.1
django-extensions==2.1.4
requests==2.21.0
model_mommy == 1.6.0
selenium == 3.141.0
django-filter == 2.1.0
